Started new abbreviations
